Change coordinate system to be sane
To make the coordinates compatible with the format we are supposed to be
using for class (see the link in the syllabus), I flipped the boards y
axis so (0,0) is closest to piece at (0,1), which is game square 1
instead of 4.  This will make the internal calculations easier.  I also
changed your loop counters from i and j to x and y respectively.  Clear
variable names make life easier for your team members :)