Create board class and delegate responsibilities
I created a board class which the gui will delegate responsibility of
maintaining itself and the pieces.

In the future the board object should be able to handle things such as
telling whether a given move is valid based upon a given input.
Additionally, it will handle the execution of moves. issued by the gui.

The screen object will get passed into the board object and likewise to
the piece object per the idea that an object should know how to draw
itself, it just needs to know where.