The computer can jump pieces!
So far only black characters can jump as the computer. White is human
atm. hsaloiye added the get_color method earlier to piece.py, just
needed to make the change in Github so that it didn't crash anyone's
program. Also, gui has some logic in it to make it alternate between
human and computer turns, but I'm expecting it to be changed in the
future...just needed it done for testing purposes. Let me know if there
are any questions.