The computer can do multiple jumps
I added some logic to make the computer do more then one jump without
making an additional move. I believe it works, but someone else may feel
free to test it and make sure (plus I wasn't able to set it up for a
triple jump no matter how hard I tried, so that needs to be tested as
well, but Im pretty sure it will work). Again, this is only for when
human is white and computer is black. I will be implementing how to make
the computer play as the white player next.