User Double Jumps
For now, the way the user enters a double jump is by entering three separate board locations.
These are validated and then executed.  There is now a function called human_double that takes in 3 coordinates
and a function called check_jump which ensures the user is completing two separate valid jumps.