Added "king-ing" to computer pieces
The computer can now act as a king piece. A peculiar thing happens once
a king is made with the computer though...it seems to only focus on
moving that piece and will just keep moving it back and forth and will
not move any other piece. Haven't looked into it yet, but more than
likely due to the way we are only telling it to act on the first state
in the list.