Added "king-ing" to computer pieces
The computer can now move around the board as a king piece. However,
once the computer gets a king, that is pretty much the only piece it
will move unless there is a jump that another piece can take. Probably
due to how we are telling it to update the board.