Extensive changes
I added a method to validate whether or not the human player has any
moves. I made a call to it in gui.py. At the moment, it still needs to
get user input before it says the user is out of moves...will probably
have to relocate it to elsewhere.

In board.py I added the method mentioned above. I also noticed a problem
where a computer king piece would jump off of a valid spot on the board
to make a jump and then would come back onto the board at a later time.
I fixed that, but it's really ugly. Also, I believe I took care of the
issue where it sometimes says that the human has a valid jump when in
fact they dont.

Again, please make sure to test these out and communicate whether or not
you find any bugs.