Fixed a small bug with board
I realized that at one point when Im checking for jumps, I accidently
checked for a valid move instead of the jump, so the forwardLeft
direction jump was not acting as it should. Changed it to computer_jump
and it appears to be working as it should.

Also with gui.py I just commented out the section where it is trying to
determine if clicking occurs. I kept clicking on the board and it would
cause my program to crash. We can un-comment it once that has been
fixed.

Is there any way we can get rid of that dialogue box that just started
popping up?