Computer can no longer keep jumping if king
I believe I removed the capability of the computer to continue jumping
if it has just become a king from jumping that turn. I was only able to
test it with one specific sequence and it appeared to work in that case,
but please test with other cases if possible. (I was having trouble
getting the pieces to line up.)