Made some slight code changes
Min/max still works. Added some code to possibly be used/passed into the
eval function. Also, made the my_max function return a random move if
the max move score has a tie for its value. Same for min.

Noticed that there is a bug in the human jumps when kinging. Once the
human's piece gets to the king row and becomes a king, if there is
another jump available it wants to take it. Will need to look into how
to fix that.